backend/requirements.txt
fastapi
uvicorn[standard]
supabase
openai
httpx
python-dotenv
